First attempt at DowndraftMassFlux (fails) + removing duplicate field downdraft_origin_level - only State version remains

+++ GEOSagcm_GridComp/GEOSphysics_GridComp/GEOSmoist_GridComp/pyMoist/pyMoist/convection/GF_2020/cumulus_parameterization/state.py
@@ -271,7 +271,7 @@ class GF2020CumulusParameterizationState(State):
                 "dims": [X_DIM, Y_DIM, "plumes"],
                 "units": "?",
                 "intent": "?",
-                "dtype": Float,
+                "dtype": Int,
             }
         )
         lcl_level: Quantity = dataclasses.field(
